fix: Changed directories
